pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "swoosh-cli"
version = "0.1.0"
description = "All-in-one CLI for Git workflow automation: init, commit, release, deploy, sync"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "MNametissa"}]
keywords = ["git", "github", "cli", "automation", "deploy", "ci-cd"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
]

dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
    "pyyaml>=6.0",
    "toml>=0.10.0",
    "paramiko>=3.0.0",
]

[project.scripts]
swoosh = "swoosh.cli:app"

[project.urls]
Homepage = "https://github.com/MNametissa/swoosh"
Repository = "https://github.com/MNametissa/swoosh"
